requirements.txt
python-dotenv==0.20.0
aiogram==2.20
pandas==1.4.2
fastapi==0.75.2
aiohttp==3.8.1
websockets==10.3
pydantic==1.9.1
uvicorn==0.17.6
openpyxl==3.0.10
asyncpg==0.25.0
pymorphy2==0.9.1
scikit-learn==1.1.1
ujson==5.3.0
pymongo==4.1.1
motor==3.0.0